requirements.txt
amqp==5.0.3
asgiref==3.2.10
billiard==3.6.3.0
boto3==1.17.109
botocore==1.20.109
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==3.1
django-cors-headers==3.6.0
django-dbbackup==3.3.0
django-extensions==3.1.1
django-filter==2.4.0
django-storages==1.11.1
django-utils-six==2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==5.0.2
MarkupSafe==1.1.1
moviepy==1.0.3
numpy==1.21.1
packaging==20.8
Pillow==8.0.1
proglog==0.1.9
prompt-toolkit==3.0.14
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.5
razorpay==1.2.0
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.4.2
six==1.15.0
sqlparse==0.4.1
tqdm==4.62.0
twilio==6.54.0
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
